moved 2 functions in wrong module
