+++ src/spatialdata/transformations/__init__.py
@@ -1,6 +1,8 @@
 from spatialdata.transformations.operations import (
+    align_elements_using_landmarks,
     get_transformation,
     get_transformation_between_coordinate_systems,
+    get_transformation_between_landmarks,
     remove_transformation,
     set_transformation,
 )
@@ -26,4 +28,6 @@ __all__ = [
     "set_transformation",
     "remove_transformation",
     "get_transformation_between_coordinate_systems",
+    "get_transformation_between_landmarks",
+    "align_elements_using_landmarks",
 ]
